requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.9.0+cpu 
torchvision==0.10.0+cpu
boto3
ffmpeg
Pillow
requests
